Remove Rustfmt from CI and include stable in CI
Rustfmt was not behaving correctly for the doctest code. I would have
preferred for it to use the "standard" Rust style but I would have
settled for anything remotely sensible looking. Unfortunately, that
wasn't the case and I couldn't figure out how to fix it.

I'll just keep manually trying to enforce the style and watch out for
any updates that may fix the issue.

See !26 for slightly more information.