Implement most of the non-callback AIO
I'm not a fan of the name and I don't know how I want to implement the
`wait` function, but most of it is there and the logic is _much_
cleaner.

I also changed the AIO trait to take `&mut self` for all of the
functions. It makes things cleaner in the non-callback API and the
callback API can just depend on cloning. I may switch that up again if I
decide to enable the cloning of the non-callback version but I think
this works well for now.