Link to a newer version of NNG
Banyan actually managed to consistently trigger a bug that caused the
AIO operations to block forever instead of returning instantly. This
updates the crate to utilize a version that does not have that bug.

This branch will probably be good to die as soon as Banyan is updated to use the `nng-sys 1.2.0-rc.x` crate.