Change Mutex+Arc to RwLock in pipe notify
The change from Mutex to RwLock is fairly obvious: in all (sane?)
implementations, a large number of readers with very few writes will
perform better with a RwLock. A properly designed program should only
set the callback a small number of times, likely only a single time
right after the socket is created.

The change to drop the Arc is less obvious and less easily justified.
The strongest bit of support I have for it is that is simplifies the
logic of the code. However, it also removes an extra atomic operation
that happens each time the callback is run, which may or may not be
expensive (I don't know and it probably depends on the system). Given
the rate that the callbacks will likely be happening and the rate at
which changing the callbacks should be happening, this kind of just
makes sense?